requirements.txt
matplotlib==3.10.8
pandas==3.0.2
Pillow==12.2.0
plotly==6.0.0
openpyxl